+++ src/azure-cli/azure/cli/command_modules/vm/_template_builder.py
@@ -624,7 +624,7 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
                         single_placement_group=None, platform_fault_domain_count=None, custom_data=None,
                         secrets=None, license_type=None, zones=None, priority=None, eviction_policy=None,
                         application_security_groups=None, ultra_ssd_enabled=None, proximity_placement_group=None,
-                        terminate_notification=None, terminate_notification_time=None):
+                        terminate_notification_time=None):
 
     # Build IP configuration
     ip_configuration = {
@@ -796,11 +796,11 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
     if proximity_placement_group:
         vmss_properties['proximityPlacementGroup'] = {'id': proximity_placement_group}
 
-    if terminate_notification is not None:
+    if terminate_notification_time is not None:
         scheduled_events_profile = {
             'terminateNotificationProfile': {
                 'notBeforeTimeout': terminate_notification_time,
-                'enable': terminate_notification
+                'enable': 'true'
             }
         }
         vmss_properties['virtualMachineProfile']['scheduledEventsProfile'] = scheduled_events_profile